refactor: extract thread and async api
